defconfig: enable power-efficient workqueues
This commit enables power-efficient workqueues, introduced in commit cee22a1, which were not enabled in upstream kernel configs.

Power-efficient workqueues are marked as unbound, allowing jobs to run on any CPU. This improves power efficiency by letting the scheduler decide where to run tasks.

This change enables the feature for ARM64 platforms, the main target for this improvement.

Change-Id: I6eefe1e967701c62e557c0db7820c281b1349d08
Signed-off-by: fadlyas07 <mhmmdfdlyas@gmail.com>
Signed-off-by: Lseything <lseything@proton.me>